+++ benches/hierarchy/storage/mod.rs
@@ -15,7 +15,7 @@ use criterion::{black_box, Bencher, Criterion};
 use kvproto::kvrpcpb::Context;
 use test_storage::SyncTestStorageBuilder;
 use test_util::KvGenerator;
-use tikv::storage::engine::Engine;
+use tikv::storage::kv::Engine;
 
 use tikv::engine::CF_DEFAULT;
 use tikv::storage::{Key, Mutation};
