+++ benches/hierarchy/txn/mod.rs
@@ -14,7 +14,7 @@
 use criterion::{black_box, Bencher, Criterion};
 use kvproto::kvrpcpb::Context;
 use test_util::KvGenerator;
-use tikv::storage::engine::Engine;
+use tikv::storage::kv::Engine;
 use tikv::storage::mvcc::MvccTxn;
 use tikv::storage::{Key, Mutation, Options};
