+++ src/coprocessor/dag/executor/metrics.rs
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::storage::engine::Statistics;
+use crate::storage::kv::Statistics;
 use prometheus::local::LocalIntCounterVec;
 
 /// `ExecutorMetrics` is metrics collected from executors group by request.
