+++ src/coprocessor/local_metrics.rs
@@ -16,7 +16,7 @@ use std::mem;
 use crate::coprocessor::dag::executor::ExecutorMetrics;
 use crate::coprocessor::metrics::*;
 use crate::pd::PdTask;
-use crate::storage::engine::{FlowStatistics, Statistics};
+use crate::storage::kv::{FlowStatistics, Statistics};
 use crate::util::collections::HashMap;
 use crate::util::worker::FutureScheduler;
 use prometheus::local::{LocalHistogramVec, LocalIntCounterVec};
