+++ src/server/errors.rs
@@ -24,7 +24,7 @@ use protobuf::ProtobufError;
 use super::snap::Task as SnapTask;
 use crate::pd::Error as PdError;
 use crate::raftstore::Error as RaftServerError;
-use crate::storage::engine::Error as EngineError;
+use crate::storage::kv::Error as EngineError;
 use crate::storage::Error as StorageError;
 use crate::util::codec::Error as CodecError;
 use crate::util::worker::ScheduleError;
