+++ src/storage/mvcc/reader/scanner/backward.rs
@@ -16,7 +16,7 @@ use std::cmp::Ordering;
 use kvproto::kvrpcpb::IsolationLevel;
 
 use crate::engine::CF_DEFAULT;
-use crate::storage::engine::SEEK_BOUND;
+use crate::storage::kv::SEEK_BOUND;
 use crate::storage::mvcc::write::{Write, WriteType};
 use crate::storage::mvcc::Result;
 use crate::storage::{Cursor, Key, Lock, Snapshot, Statistics, Value};
