+++ src/storage/txn/mod.rs
@@ -28,7 +28,7 @@ use crate::util::escape;
 quick_error! {
     #[derive(Debug)]
     pub enum Error {
-        Engine(err: crate::storage::engine::Error) {
+        Engine(err: crate::storage::kv::Error) {
             from()
             cause(err)
             description(err.description())
